+++ org.school.rumorspread/src/main/java/org/school/rumorspread/RumorSpreadVertexValue.java
@@ -5,26 +5,26 @@ import java.util.Iterator;
 
 import org.apache.commons.lang.StringUtils;
 import org.apache.giraph.utils.ArrayListWritable;
-import org.apache.hadoop.io.DoubleWritable;
+import org.apache.hadoop.io.FloatWritable;
 
-public class RumorSpreadVertexValue extends ArrayListWritable<DoubleWritable> {
+public class RumorSpreadVertexValue extends ArrayListWritable<FloatWritable> {
 
 	private static final long serialVersionUID = 227809916468865164L;
 
 	public RumorSpreadVertexValue() {
 	}
 	
-	public RumorSpreadVertexValue(ArrayList<DoubleWritable> values) {
+	public RumorSpreadVertexValue(ArrayList<FloatWritable> values) {
 		this.clear();
 		this.addAll(values);
 	}
 	
 	@Override
 	public void setClass() {
-		setClass(DoubleWritable.class);
+		setClass(FloatWritable.class);
 	}
 
-	public DoubleWritable getLastValue() {
+	public FloatWritable getLastValue() {
 		return this.get(this.size() - 1);
 	}
 	
@@ -35,7 +35,7 @@ public class RumorSpreadVertexValue extends ArrayListWritable<DoubleWritable> {
 	public double getValuesSum(int t) {
 		double sum = 0.0;
 		
-		Iterator<DoubleWritable> itr = this.iterator();
+		Iterator<FloatWritable> itr = this.iterator();
 		for (int i = 0; i < t && i < this.size() && itr.hasNext(); i++) {
 			sum +=  itr.next().get();
 		}
@@ -46,7 +46,7 @@ public class RumorSpreadVertexValue extends ArrayListWritable<DoubleWritable> {
 	public String[] toStrings() {
 		String[] strings = new String[this.size()];
 		
-		Iterator<DoubleWritable> itr = this.iterator();
+		Iterator<FloatWritable> itr = this.iterator();
 		for (int i = 0; i < this.size() && itr.hasNext(); i++) {
 			strings[i] = itr.next().toString();
 		}
